delete podAnalysis.go from Abe's PR
